Positron ipywidgets listen to results messages and show plots pane (#4195)
Addresses #4181 


This PR does three things

1. Listens to `LanguageRuntimeMessageType.Result` messages in addition
to (the existing) `LanguageRuntimeMessageType.Output` messages. jupyter
messages with widgets attached can come over both (e.g. leaflet plots
come across on `.Result`.
2. Adds some extra steps for cleaning up disposables related to the
`PositronIPyWidgetsService` listening to a given session. Now we check
if there are existing listeners for the given session and make sure to
dispose them, where previously it was possible to leak when the
application was reloaded.
3. Makes sure the plots pane is shown when an iPyWidget is rendered.
Previously if the plots pane was not visible the user might not see if
the plot was shown.

<img width="1002" alt="image"
src="https://github.com/user-attachments/assets/bf1f4bc3-aa0f-4343-9031-5900c9cf41b3">


### Notes
- If we want to remove the work for 2, I'm fine with that. It's
defensive but maybe it complicates the diff.

### QA Notes

Running the following leaflet code in a Python script should open the
viewer pane (even if it was hidden) and show a map.


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->