Data Explorer: Add size limit for object dtype type inference for pandas (#4296)
This is a performance optimization for large data frames where we have
been using `pandas.api.types.infer_dtype` to exhaustively determine a
data type for `object` dtype columns, but for very large data frames,
this can result in a large delay in calls to `get_schema`. This puts in
a cap at 1 million elements for doing this type inference, beyond which
(if inconclusive -- e.g. all null values) we will show "unknown" in the
UI. But these scenarios will be pretty infrequent, so I think it's a
reasonable tradeoff to improve the > 99% of scenarios where the type can
be well-determined by looking at this many data elements.