Added testrail upload for PR & Merge to main (#4308)
### Intent
Address #4146
Allow PRs and merge-to-main smoke test runs to upload to TestRail

### Approach

Instead of trying to parse out if it was a PR or merge-to-main.. just treat them both the same. The branch name should allow for delineation.

Also made slight change to nightly smoke run title to clarify as well.

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->