Use custom tools for universal binary reconciliation
`filesToSkip` no longer works for us. It is not consulted for x64-only files,
which should be ignored via `x64ArchFiles`, which only accepts a single glob
pattern. So we now have our own procedure to stash away such files and move them
back into the universal bundle at the end.

We also create a new set of files to reconcile following
this upstream change:
https://github.com/microsoft/vscode/commit/f6f90e016301f288ac5b36936d15ff8af9d70923

Localised strings are now stored in an array and the strings are replaced at
build time by references to this array. Somehow this array differs across arch
distributions, so we fix that up by copying from the relevant files from the x64
dist to the arm64 one.