Retry logic awaiting line advance for fast R line execution test (#4363)
Fast execution test was passing on OSX and Windows but failing on Ubuntu
executors.

This code implements multi-level retry logic to fix Ubuntu. It tracks
the current line in the editor and does not send a new "Control-Enter"
until it detects an advance. It very rapidly retries when trying to
determine the current line. No sleeping is employed because we want to
move through the lines as fast as possible.

### QA Notes

All smoke tests pass.

---------

Signed-off-by: Christopher Mead <chrisrmead@comcast.net>
Co-authored-by: Nick Strayer <nick.strayer@posit.co>