Support file download links in webviews (#4377)
Addresses #4087 
Adds logic for intercepting clicks on links for downloading assets (via
`blob:` and `data"` schemes).
This adds support for downloading plots etc from a interactive plotting
libraries like altair.

Previously for altair plots we simply hid the actions menu because it
was unfunctional. Now we re-enable it and show the svg and png download
options.
<img width="400" alt="image"
src="https://github.com/user-attachments/assets/d69ef6e3-d959-4f88-af83-ec11b12ac4c3">

<img width="1076" alt="image"
src="https://github.com/user-attachments/assets/2e9b1ffd-2934-41e6-a3e5-ef5cbcbe35e9">
<img width="1054" alt="image"
src="https://github.com/user-attachments/assets/4b33df56-00eb-4edc-ab82-cdbfc2c58b29">

(We still hide the other options because their link types still dont
open - and don't in other iframe-based environments like jupyter
notebooks, either.)

Right now the only library that I'm sure we have enabled downloading for
is Altair/vega, but in theory this is a common pattern.

### Implementation
We take inspiration from the `webviewPreloads.ts` file for the logic.
Luckily ours is a little simpler because we don't have to deal with the
context of multiple outputs and focus changing.

If the user clicks a blob url then we fetch the blob and convert to a
data url, if it's a dataurl we just send it through via a new message
type `PositronDownloadMessage` that is intercepted by a message watcher
on the webview.

### Future work
Currently, we handle a limited set of webview actions that affect
external elements. This can be expanded in the future to include more
utilities, such as opening links in the preview panel.

Also these patches are only used in the plots pane webviews currently,
but would probably be useful for all instances of the preview or plots
pane so things that use the `UiFrontendEvent.ShowHtmlFile` comm are also
given the abilities.

---------

Signed-off-by: Nick Strayer <nick.strayer@posit.co>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>