Fix possible race condition when rendering webview plots (#4461)
This is an attempt to fix the failing smoke tests for IPyWidget plots
(#4448).

I suspect the issue is that during the initial `claim` of the overlay
webview, the render message is sent before the webview preloads script
is fully initialized. This PR adds an `onDidInitialize` event to the
notebook output webview, and delays sending the render messages until
the event is fired, which is closer to how the backlayer webview works.

I also cleaned up a few other minor things that I thought were related
while investigating.

### QA Notes

The full tests CI should pass (ideally more than once).