Open file urls in external browser (#4468)
This change causes certain URLs to HTML files to open in an external
browser rather than opening up the HTML file in Positron's editor,
addressing a regression from #4151.

In the aforementioned change, we added some machinery that automatically
transforms and proxies URLs that arrive from language runtimes. This is
intended for http URLs, but also runs for file URLs, resulting in those
files being opened in the editor instead of an external browser.
Probably OK if e.g. a runtime asks us to show a `.cpp` file, but not
cool for HTML content.

The fix is to check for this situation specifically; when we're handed a
file URI that looks like HTML, we start a proxy for it and then open it
externally.

I'm not sure it's always right to open these in an external browser
(maybe sometimes we'll want the Viewer pane?) but it's what people
expect for scenarios where this is broken now, so we can start here and
see how it feels.

Addresses https://github.com/posit-dev/positron/issues/4297.

### QA Notes

- It's worth double checking a couple of the scenarios from #4151 to
make sure all that good stuff still works.
- This change is also intended to work on Workbench and on remote SSH
scenarios where we can't just load the file URL directly since it's on
the remote host.