move `package.json` merging to `postinstall.js` script
On a clean machine, the `node_modules` folder is not yet created, so the `gulp` and `gulp-merge-json` dependencies are not yet available prior to `yarn install` and thus not yet available in the `preinstall.js` script.

As such, we need to move the `package.json` merging to the `postinstall.js` script, which is executed after `yarn install` has completed in the top-level directory and where some additional per-directory install handling occurs.

In particular, the `package.json` merging must occur AFTER `yarn install` has been run in the `build` directory, so that the `gulp` and `gulp-merge-json` dependencies are available for use.

To test if the package.json merging is still working:

1. make a change to `remote/package.json`  and/or `remote/web/package.json`
2. delete the `node_modules` directories in `build` and `remote/reh-web`
3. run `yarn` in the top-level directory
4. no errors should occur and the `node_modules` directories should be recreated, and `remote/reh-web/package.json` should have the changes from `remote/package.json` and `remote/web/package.json` merged into it