Detect hyperlinks surrounded by angle brackets (#4491)
This change fixes a problem wherein links like `Look up docs at
<http://example.com/topic>` do not work in the console because the
closing `>` is interpreted as part of the link. The link detection is
currently a simple regex, so the only actual fix here is just adding `>`
to the list of characters that are considered to terminate a link.

However, the change looks bigger because I also moved the link detector
into its own file and added some unit tests to it so that we can make
this more robust.

Addresses https://github.com/posit-dev/positron/issues/4413.

### QA Notes

There are good repro cases in the bug! This only touches console-related
code.