Await default interpreter startup before starting a new one (#4500)
Occasionally we see startup crashes in tests that have been difficult to
debug. One theory as to why this is happening is that the test might be
trying to start the test interpreter at the same time the default
interpreter is still starting up. This PR adds code to wait for the
default interpreter (or definitively no interpreter) to start up before
initiating the process of starting a new one.

### QA Notes

All smoke tests should pass.