Holoview extension notification (#4503)
Addresses #4365 

This PR adds support for loading non-bokeh-based extensions for hvplot. 

It does this by
- Adding a new ui comm message: `holoviz_extension_load` 
- Adding a wrapper around the hvplots extension loading method that
sends out the new extension load message
- Forwarding the comm message through to the runtime client events
- Reacting to the message in the holoviz service to dump the current set
of stored extension messages to make way for the new ones for the loaded
extension.

The PR also fixes an error that often occurs in the console when we
replay of multiple messages that render something to the dom. The error
was caused by the fact that every successive rendering message was
wiping out the previous render results because of reusing the container.
Now a new render container is provided for each output.

### QA Notes

You should be able to run the following code in the console and a plot
should appear in the plots pane.

```python
import hvplot.pandas
import pandas as pd
hvplot.extension('plotly')
pd.DataFrame(dict(x=[1,2,3], y=[4,5,6])).hvplot.scatter(x="x", y="y")
```
<img width="1354" alt="image"
src="https://github.com/user-attachments/assets/d9267a5b-7c98-418a-a98a-d10341767732">

It should not matter if you run all the lines in a single execution or
line-by-line.

---------

Signed-off-by: Nick Strayer <nick.strayer@posit.co>
Co-authored-by: seem <mwlorgat@gmail.com>