Clean up smoke test depencies (move all to qa-example-content) (#4535)
Moving all Python dependencies other than ones that are part of the test
infrastructure to
[qa-example-content](https://github.com/posit-dev/qa-example-content).

This is in an effort to simplify the decision making process when adding
new dependencies for smoke tests. Python dependencies go in
[requirements.txt](https://github.com/posit-dev/qa-example-content/blob/main/requirements.txt)
and R dependencies go in
[DESCRIPTION](https://github.com/posit-dev/qa-example-content/blob/main/DESCRIPTION).

The dependencies are kept in a separate repo because qa-example-content
has all the same requirements and we want to only manage one list of
dependencies for R and one for Python.

The only Python dependencies which remain in Positron workflows for
smoke tests are:
* ipykernel - needed for any Python testing
* trcli - test rail cli.  needed to publish test results. 

### QA Notes

All smoke tests should pass.