python: add to bokeh testing +windows (#4533)
re: #4397

patches bokeh to not allow writing files to positron extension

A few other stragglers-- the [Python nightly tests are failing for bokeh
stuff](https://github.com/posit-dev/positron/actions/workflows/positron-python-nightly.yml),
likely due to the fact `bokeh` is not being installed. AND I am turning
the Windows tests back on, they seem to be running on my machine but if
there is other context I'm missing happy to rework tests in a different
way!