Add "run selection" command to command palette, context menu, toolbar menu (#4590)
Addresses #2084 

In #2068 we added access to the command
`workbench.action.positronConsole.executeCode` for Python to:

- the context menu (right click menu)
- the command palette
- the toolbar menu

This PR adds access to the same command in the same places for R.

### QA Notes

With this PR, you can now use the command
`workbench.action.positronConsole.executeCode` (i.e. run your current
line or selection in the console) with any of those three UI gestures.
It should look fairly similar to what you see with Python files,
although the context menu for Python is a little different.