Data Explorer: Use concurrent.futures background job queue to return get_column_profiles results asynchronously (#4326)
As discussed in #4300, comm requests have to be served in the order they
are received. Since the data explorer has some request types, like
`get_column_profiles` that can grow very expensive for large datasets,
this blocks "fast" requests like `get_data_values` from executing while
these "slow" requests are pending.

This change adds a `return_column_profiles` frontend method that allows
these requests to be fulfilled in the background, allowing fast requests
to be served unimpeded. This resolves the immediate performance issue
that the data explorer is facing, so this pattern can be refined over
time as we have need to do other asynchronous / background request
handling in the kernels.

---------

Signed-off-by: Daniel Falbel <dfalbel@gmail.com>
Co-authored-by: Daniel Falbel <dfalbel@gmail.com>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>
Co-authored-by: Chris Mead <chris.mead@posit.co>