Plots in an editor (#4364)
Creates a new editor that takes a plot id as input. The Plots view
has a button that can open the plot in an editor.

The comm has been refactored to be shared among plot clients
and will be disposed of when all clients are disposed.

The entire feature is hidden behind an experimental flag. It is in the
Application > Experimental category in the preferences.

Signed-off-by: Tim Mok <timtmok@users.noreply.github.com>
Co-authored-by: Nick Strayer <nick.strayer@posit.co>