Data Explorer: Improve performance of summary histograms in Python (#4652)
For large datasets, passing a list of bin edges to `numpy.histogram`
results in O(n log n) work which scales poorly. Since we always compute
histograms with uniform bin edges, this adapts the logic to recompute
the histogram when the Freedman-Diaconis rule yields a too fine-grained
histogram, but to remain O(n) in time complexity with the size of the
dataset.