Fix array of args for R task `ShellExecution` (#4650)
Addresses #4649 

The combination of the `vscode.ShellQuoting.Strong` for the code with
spaces and the R args being passed in as one string instead of an array
made this unhappy. I believe I introduced this regression in #4576.

### QA Notes

The "R: Install R Package and Restart R" command should work again,
along with the other R package commands.

It would be good to have tests that run these commands (just executing
the commands directly would be good enough) in an R package test
environment and check for no errors.