Use arm64 builds of Ark on Linux (#4711)
As a prerequisite to making arm64 builds of Positron, we need Positron
to bundle arm64 builds of ark when building on arm64 Linux. This change
does just that.

Ark is already producing arm64 builds of Linux, e.g.
https://github.com/posit-dev/ark/releases/tag/0.1.136.

### QA Notes

N/A; this does not currently impact any shipping releases since none are
made on arm64.