Merge 1.93 from upstream (#4655)
This change includes two large payloads:

- All the changes from Code OSS 1.92.0 -> 1.93.0
- All of the Posit Workbench modifications from their fork of VS Code

A lot of hand-editing needed to happen here since our compressed history
means we have no common baseline to use for a commit-based merge. The
methodology used was roughly:

1. Merge 1.93.0 into vscode-server and test against Workbench; make
fixes/tweaks as needed
2. Merge 1.93.0 into Positron (using our usual merge strategy)
3. Copy all the code files (which now have a common 1.93.0 baseline)
from vscode-server => Positron
4. Go through the git view for each one and reconcile the Positron and
PWB change hunks, discarding the edits that dropped Positron blocks,
staging the edits that added PWB blocks, and manually resolving the
dozen or so places where we both edited the same logic
5. Commit and continue typical upstream merge process

This was a one-shot merge process that won't need to be repeated since
next time we will have a common baseline (a tag from vscode-server
representing what we merged last time).

Nothing too crazy in here. One of the more annoying parts was getting
the `rstudio.rstudio-workbench` built-in extension set up so that it
gets bundled for the `reh-web` build but ignored in other contexts.

### QA Notes

This contains a lot of churn introduced from two different sources, so
some testing is indicated pre-merge. Here are some binaries you can
test:

https://github.com/posit-dev/positron-builds/actions/runs/10821973170

---------

Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>