tests: composite action for caching, remove fallback restore keys (#4709)
This update refactors the caching steps in the merge-to-main and
full-test-suite workflows by consolidating them into a GitHub composite
action. This change enhances maintainability by reducing code
duplication. Additionally, fallback restore keys have been removed,
ensuring that cache restoration will only occur when an exact match to
the lock file is found, improving cache reliability and avoiding stale
dependencies.

Update:
In this PR, after I moved the restore/save cache logic into composite
actions, I noticed the cache was being restored correctly. However, the
cache hit always returned false. It seems the downstream job doesn’t
have access to the output from the step when it’s part of a composite
action, which is disappointing.

Instead of keeping the the save/restore process split, I reverted to
using the default cache action (which combines it into one) and saves
the cache as a post step when everything completes. The drawback is that
if there’s a failure (such as a test error), the save cache step won’t
execute.

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>
Co-authored-by: Jon Vanausdeln <jonv@live.com>
Co-authored-by: sharon <sharon-wang@users.noreply.github.com>
Co-authored-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>
Co-authored-by: Jonathan <jonathan@rstudio.com>
Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>
Co-authored-by: Christopher Mead <chrisrmead@comcast.net>