Move the module to peerDependencies
Goal is to prevent @vscode/windows-registry from ending up in the `node_modules` directory of the built positron-r extension.

Its presence there is breaking the macOS universal app because a generated Makefile is different on arm64 vs x86. But the root problem is that the module just shouldn't be there at all!