Improve keyboard navigation for data viewer filter dialogs (#4620)
- If focus is on the column drop down, then pressing a key will
immediately open the search sub-dialog and use that key as an initial
search input.
- Updating the search string will always place the cursor on the first
available item
- Hitting "ArrowDown" or "Tab" from within the search input will shift
focus to the list and allow keyboard navigation as expected
- Hitting "Enter" or "Space" when navigating the list will confirm the
selection

I've also added a new style to the filter widgets to use monospace font
for the values. This makes them more legible in my opinion, particularly
with values which contain "l" "1" and "I".