Update logs in the `jupyter-adapter` extension (#4845)
Addresses #4275 

This PR makes changes to all the output channels generated by the
`jupyter-adapter` extension.

The "main" Jupyter Adapter output channel is now a real log, with
timestamps and log level annotations:

![Screenshot 2024-09-28 at 4 51
07 PM](https://github.com/user-attachments/assets/28c48414-f7ee-4da5-9ebb-eeaa96d892ac)

The existing "Console: X.Y.Z" channel now only has messages generated
from Positron, not the kernel itself. This is now also a real log and is
the part that directly addresses #4275:

![Screenshot 2024-09-28 at 4 51
26 PM](https://github.com/user-attachments/assets/9e1b1345-4001-44d6-b43a-d2b9e092027c)

Notice the new "error" log level here; you can now set the log level in
these first two output channels and it will update moving forward.

These first two output channels are generated in the same way most of
the time (we write to the "Console" channel if that kernel is up but to
the "Jupyter Adapter" channel if not) so we have to make changes to them
together; this seems good rather than bad to me and overall both are now
improved.

There is a new "Kernel: X.Y.Z" channel which still uses the raw output
channel that we created back in #1589. It is not a real log and does not
have timestamps, unless the kernel log generated those. You cannot
change the log level via the normal method for this one:

![Screenshot 2024-09-28 at 4 51
38 PM](https://github.com/user-attachments/assets/2df12fb3-7cc5-4aab-b49b-933fc3ecdc7a)

It is kind of a bummer to now have _three_ output channels generated
from this one extension but IMO the tradeoffs are worth it.


### QA Notes

You'll now see timestamps and log level annotations in the "Jupyter
Adapter" and "Console: X.Y.Z" output channels, and there is a new
"Kernel: X.Y.Z" channel.