Memoise findCurrentRBinary()
This often gets called outside of (and before) the main R installation discovery effort, when validating the metadata of a previously-used R runtime.

We don't claim to watch for changes to which R installation is current. Therefore, we only need to discover the current R once.

I'm mostly doing this to reduce noise in the positron-r output channel.