notebookCreate test: fix test setup (#4922)
### Intent
Noticed two issues with this test. First, the tests within each describe
block are not independent and rely on the first one. So I re-organized
the before/beforeEach blocks so each test can standalone. The 2nd issue
is that if the 2nd tests attempt to retry they will potentially fail
because there will be duplicate matching markdown text from the first
attempt.

### Approach
1. New notebook for each test
2. Random text used to verify h2 markdown